+++ src/main/java/repast/model/agrosuccess/AgroSuccessContextBuilder.java
@@ -2,10 +2,8 @@ package repast.model.agrosuccess;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
 import org.apache.commons.configuration.ConfigurationException;
 import org.apache.log4j.Logger;
 import org.neo4j.graphdb.GraphDatabaseService;
@@ -36,6 +34,8 @@ import me.ajlane.neo4j.EmbeddedGraphInstance;
 import repast.model.agrosuccess.AgroSuccessCodeAliases.Lct;
 import repast.model.agrosuccess.empirical.SiteRasterData;
 import repast.model.agrosuccess.empirical.SiteWindData;
+import repast.model.agrosuccess.params.EnvrModelParams;
+import repast.model.agrosuccess.params.ModelParamsRepastParser;
 import repast.model.agrosuccess.empirical.SiteAllData;
 import repast.model.agrosuccess.empirical.SiteAllDataFactory;
 import repast.model.agrosuccess.empirical.SiteClimateData;
@@ -49,7 +49,6 @@ import repast.simphony.dataLoader.ContextBuilder;
 import repast.simphony.engine.environment.RunEnvironment;
 import repast.simphony.engine.schedule.ISchedule;
 import repast.simphony.engine.schedule.ScheduleParameters;
-import repast.simphony.parameter.IllegalParameterException;
 import repast.simphony.parameter.Parameters;
 import repast.simphony.space.grid.GridBuilderParameters;
 import repast.simphony.space.grid.SimpleGridAdder;
@@ -71,48 +70,6 @@ public class AgroSuccessContextBuilder implements ContextBuilder<Object> {
 
   final static Logger logger = Logger.getLogger(AgroSuccessContextBuilder.class);
 
-  // pseudo-agents modify GridValueLayer-s
-  SoilMoistureCalculator soilMoistureCalculator;
-
-  GridValueLayer soilMoisture; // name = "soil moisture"
-  GridValueLayer soilTypeMap; // see SoilMoistureCalculator, name = "soil"
-  GridValueLayer landCoverTypeMap; // name = "lct"
-  GridValueLayer succession;
-
-  GridValueLayer slopeMap; // name = "slope"
-  GridValueLayer aspect; // name = "aspect"
-
-  GridValueLayer pineSeeds; // name = "pine seeds"
-  GridValueLayer oakSeeds; // name = "oak seeds"
-  GridValueLayer deciduousSeeds; // name = "deciduous seeds"
-
-  GridValueLayer timeInState;
-  GridValueLayer deltaD;
-  GridValueLayer deltaT;
-
-  public void endMethod(Context<Object> context, RecordWriter<Lct, Double> lctWriter) {
-    logger.info("End of the simulation");
-    for (Object graph : context.getObjects(EmbeddedGraphInstance.class)) {
-      ((GraphDatabaseService) graph).shutdown();
-    }
-
-    try {
-      lctWriter.flush();
-    } catch (IOException e) {
-      // TODO Auto-generated catch block
-      e.printStackTrace();
-    }
-  }
-
-  public void updateLctWriter(LctProportionAggregator aggregator,
-      RecordWriter<Lct, Double> lctWriter) {
-    lctWriter.add(aggregator.getLctProportions());
-  }
-
-  public void printLctProportion(LctProportionAggregator lctPropAgg) {
-    logger.debug(lctPropAgg.getLctProportions());
-  }
-
   @Override
   public Context<Object> build(Context<Object> context) {
 
@@ -121,8 +78,7 @@ public class AgroSuccessContextBuilder implements ContextBuilder<Object> {
     SimulationID simulationID = new SimulationID(params.getString("studySite"));
     modelCore.endAt(params.getInteger("nTicks"));
 
-    // TODO Add parameters required by ModelParamsRepastParser to parameters.xml
-    // EnvrModelParams envrModelParams = new ModelParamsRepastParser(params);
+    EnvrModelParams envrModelParams = new ModelParamsRepastParser(params);
 
     File databaseDir = new File((String) params.getValue("graphPath"), "graph.db");
     GraphDatabaseService graph = new EmbeddedGraphInstance(databaseDir.getAbsolutePath());
@@ -137,46 +93,24 @@ public class AgroSuccessContextBuilder implements ContextBuilder<Object> {
 
     initGridValueLayers(context, siteData);
 
-    SeedDisperser seedDisperser = initSeedDisperser(context, siteData, params);
+    SeedDisperser seedDisperser = initSeedDisperser(context, siteData,
+        envrModelParams.getSeedDispersalParams(), envrModelParams.getSeedViabilityParams());
     context.add(seedDisperser);
 
     SoilMoistureCalculator smCalc = initSoilMoistureCalculator(context, siteData);
     context.add(smCalc);
 
-    LcsUpdater lcsUpdater = initLcsUpdater(context, graph, params);
+    LcsUpdater lcsUpdater = initLcsUpdater(context, graph, params.getString("graphModelID"),
+        envrModelParams.getSoilMoistureParams());
     context.add(lcsUpdater);
 
     // TODO Configure land cover flammability replicate in parameters.xml
     FireManager fireManager = initFireManager(context.getValueLayer(LscapeLayer.Dem.name()),
-        (IGridValueLayer) context.getValueLayer(LscapeLayer.Lct.name()),
-        siteData, siteData, siteData, LcfReplicate.Default);
+        (IGridValueLayer) context.getValueLayer(LscapeLayer.Lct.name()), siteData, siteData,
+        siteData, LcfReplicate.Default);
     context.add(fireManager);
 
-    // StudySiteDataContainer siteDataContainer =
-    // new StudySiteDataContainer(new File(params.getValueAsString("siteDataRoot"),
-    // params.getValueAsString("studySite")));
-    // FireManager fireManager = initialiseFireManager(context,
-    // siteDataContainer.getWindDirectionProb(), siteDataContainer.getWindSpeedProb(),
-    // siteDataContainer.getGridCellPixelSize()[0], 0, 0, null);
-
-    ISchedule sche = RunEnvironment.getInstance().getCurrentSchedule();
-
-    try {
-      LctProportionAggregator lctPropAgg = new LctProportionAggregator(context.getValueLayer(LscapeLayer.Lct.name()));
-      RecordWriter<Lct, Double> lctWriter;
-      lctWriter = new EnumRecordCsvWriter<Lct, Double>(Lct.class,
-          new File("output", simulationID.toString() + "_lct-props.csv"));
-      ScheduleParameters printProps = ScheduleParameters.createRepeating(0, 1, -10);
-      sche.schedule(printProps, this, "updateLctWriter", lctPropAgg, lctWriter);
-
-      // call method at the end of the simulation run. See
-      // https://martavallejophd.wordpress.com/2012/03/26/run-a-method-at-the-end-of-the-simulation/
-      ScheduleParameters stop = ScheduleParameters.createAtEnd(ScheduleParameters.LAST_PRIORITY);
-      sche.schedule(stop, this, "endMethod", context, lctWriter);
-    } catch (IOException e) {
-      // TODO Auto-generated catch block
-      e.printStackTrace();
-    }
+    initLctReporters(context, simulationID);
 
     return context;
   }
@@ -264,39 +198,14 @@ public class AgroSuccessContextBuilder implements ContextBuilder<Object> {
    *
    * @param context Simulation context object
    * @param siteRasterData Raster data for study site
-   * @param params Repast simulation parameters (specified in {@code parameters.xml})
+   * @param sdParams Seed dispersal parameters
+   * @param svParams Seed viability parameters
    * @returns Configured SeedDisperser
    */
   private SeedDisperser initSeedDisperser(Context<Object> context, SiteRasterData siteRasterData,
-      Parameters params) {
+      SeedDispersalParams sdParams, SeedViabilityParams svParams) {
 
     double[] gridPixelSize = siteRasterData.getGridCellPixelSize();
-
-    Map<String, Double> seedParams = new HashMap<>();
-    seedParams.put("seedLifetime", 7.0);
-    seedParams.put("acornLocationParam", 3.844);
-    seedParams.put("acornScaleParam", 0.851);
-    seedParams.put("acornMaxLognormalDist", 550.0);
-    seedParams.put("windDistDecreaseParam", 5.0);
-    seedParams.put("windMinExpDist", 75.0);
-    seedParams.put("windMaxExpDist", 100.0);
-
-    for (Map.Entry<String, Double> entry : seedParams.entrySet()) {
-      try {
-        entry.setValue(params.getDouble(entry.getKey()));
-      } catch (IllegalParameterException e) {
-        logger.warn("Could not find entry for '" + entry.getKey() + "' in parameters."
-            + " Defaulting to default value " + entry.getValue());
-      }
-    }
-
-    SeedViabilityParams svParams =
-        new SeedViabilityParams(seedParams.get("seedLifetime").intValue());
-    SeedDispersalParams sdParams = new SeedDispersalParams(seedParams.get("acornLocationParam"),
-        seedParams.get("acornScaleParam"), seedParams.get("acornMaxLognormalDist"),
-        seedParams.get("windDistDecreaseParam"), seedParams.get("windMinExpDist"),
-        seedParams.get("windMaxExpDist"));
-
     SeedDisperser seedDisperser = new SpatiallyRandomSeedDisperser(gridPixelSize[0],
         gridPixelSize[1], svParams, sdParams, context);
 
@@ -324,48 +233,37 @@ public class AgroSuccessContextBuilder implements ContextBuilder<Object> {
    *
    * @param context Simulation context
    * @param graph Graph database service containing model configuration
-   * @param params Repast model parameters
+   * @param graphModelID ID of model stored in the graph
+   * @param smParams Soil moisture parameters
    * @return Configured LcsUpdater object
    */
   private LcsUpdater initLcsUpdater(Context<Object> context, GraphDatabaseService graph,
-      Parameters params) {
-
-    String modelID;
-    try {
-      modelID = params.getString("graphModelID");
-    } catch (IllegalParameterException e) {
-      logger.error("Could not find 'graphModelID' in parameters. Check configuration.");
-      throw e;
-    }
-
-    Map<String, Integer> smParams = new HashMap<>();
-    smParams.put("mesicThreshold", 500);
-    smParams.put("hydricThreshold", 1000);
-
-    for (Map.Entry<String, Integer> entry : smParams.entrySet()) {
-      try {
-        entry.setValue(params.getInteger(entry.getKey()));
-      } catch (IllegalParameterException e) {
-        logger.warn("Could not find entry for '" + entry.getKey() + "' in parameters."
-            + " Defaulting to default value " + entry.getValue());
-      }
-    }
-
-    SoilMoistureParams soilMoistureParams =
-        new SoilMoistureParams(smParams.get("mesicThreshold"), smParams.get("hydricThreshold"));
+      String graphModelID, SoilMoistureParams smParams) {
 
     EnvrStateAliasTranslator translator = new AgroSuccessEnvrStateAliasTranslator();
-    LcsTransitionMapFactory fac = new GraphBasedLcsTransitionMapFactory(graph, modelID, translator);
+    LcsTransitionMapFactory fac =
+        new GraphBasedLcsTransitionMapFactory(graph, graphModelID, translator);
     CodedLcsTransitionMap codedMap = fac.getCodedLcsTransitionMap();
     LcsUpdateDecider updateDecider = new AgroSuccessLcsUpdateDecider(codedMap);
 
-    SoilMoistureDiscretiser smDiscretiser =
-        new AgroSuccessSoilMoistureDiscretiser(soilMoistureParams);
+    SoilMoistureDiscretiser smDiscretiser = new AgroSuccessSoilMoistureDiscretiser(smParams);
 
     LcsUpdater lcsUpdater = new AgroSuccessLcsUpdater(context, updateDecider, smDiscretiser);
     return lcsUpdater;
   }
 
+  /**
+   * Create a FireManager pseudo-agent and configure it to run the fire regime in the simulation.
+   *
+   * @param demLayer Digital Elevation Model as a {@code ValueLayer}
+   * @param lctLayer Land cover type as a {@code IGridValueLayer}
+   * @param windData Site-specific wind speed and direction data
+   * @param rasterData Information about site's raster grids, used for cell size
+   * @param climateData Site-specific climate data (temperature and precipitation)
+   * @param lcfReplicate Which land cover flammability replicate to use (See Millington et al. 2009
+   *        Table 6).
+   * @return Configured FireManager
+   */
   private FireManager initFireManager(ValueLayer demLayer, IGridValueLayer lctLayer,
       SiteWindData windData, SiteRasterData rasterData, SiteClimateData climateData,
       LcfReplicate lcfReplicate) {
@@ -380,6 +278,83 @@ public class AgroSuccessContextBuilder implements ContextBuilder<Object> {
     double meanNumFires =
         12 * (climateData.getMeanAnnualTemperature() / climateData.getTotalAnnualPrecipitation());
     return new FireManager(meanNumFires, fireSpreader);
+  }
+
+  /**
+   * Create objects used for reporting land cover type proportions to disk. Make the Repast
+   * scheduler aware of their methods and configure them to run at appropriate times.
+   *
+   * @param context Simulation context
+   * @param id Simulation ID, used to name output files.
+   */
+  private void initLctReporters(Context<Object> context, SimulationID id) {
+    ISchedule sche = RunEnvironment.getInstance().getCurrentSchedule();
+    // call method at the end of the simulation run. See
+    // https://martavallejophd.wordpress.com/2012/03/26/run-a-method-at-the-end-of-the-simulation/
+    ScheduleParameters stop = ScheduleParameters.createAtEnd(ScheduleParameters.LAST_PRIORITY);
+    ScheduleParameters updateReporters = ScheduleParameters.createRepeating(0, 1, -10);
+
+    LctProportionAggregator lctPropAgg =
+        new LctProportionAggregator(context.getValueLayer(LscapeLayer.Lct.name()));
+    RecordWriter<Lct, Double> lctWriter = initLctWriter(id);
+
+    sche.schedule(updateReporters, this, "updateLctWriter", lctPropAgg, lctWriter);
+    sche.schedule(stop, this, "finaliseSimulation", context, lctWriter);
+  }
+
+  /**
+   * @param id ID of simulation model, used to generate name for output file.
+   * @return Writer object responsible for committing land cover type proportion results to disk.
+   */
+  private RecordWriter<Lct, Double> initLctWriter(SimulationID id) {
+    RecordWriter<Lct, Double> lctWriter;
+    try {
+      lctWriter = new EnumRecordCsvWriter<Lct, Double>(Lct.class,
+          new File("output", id.toString() + "_lct-props.csv"));
+    } catch (IOException e) {
+      throw new RuntimeException("Could not initialise land cover type writer.");
+    }
+    return lctWriter;
+  }
+
+
+  /**
+   * Code to run at the end of the simulation. Releases resources and reports outputs to disk.
+   *
+   * @param context Simulation concept
+   * @param lctWriter Writer used to send output to disk file.
+   */
+  public void finaliseSimulation(Context<Object> context, RecordWriter<Lct, Double> lctWriter) {
+    logger.info("End of the simulation");
+    for (Object graph : context.getObjects(EmbeddedGraphInstance.class)) {
+      ((GraphDatabaseService) graph).shutdown();
+    }
+
+    try {
+      lctWriter.flush();
+    } catch (IOException e) {
+      throw new RuntimeException("Could not flush land cover type writer");
+    }
+  }
+
+  /**
+   * Add current land cover type proportions to object which will write them to disk.
+   *
+   * @param lctAggregator Object aware of current land cover type proportions
+   * @param lctWriter Commits information about land cover type proportion to disk
+   */
+  public void updateLctWriter(LctProportionAggregator lctAggregator,
+      RecordWriter<Lct, Double> lctWriter) {
+    lctWriter.add(lctAggregator.getLctProportions());
+  }
+
+  /**
+   * Write current land cover type proportions to console.
+   *
+   * @param lctAggregator Object aware of current land cover type proportions
+   */
+  public void printLctProportion(LctProportionAggregator lctAggregator) {
+    logger.debug(lctAggregator.getLctProportions());
   }
 
 }